+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project's packages adheres to [Semantic Versioning](http://semver.org/s
 ### Fixed
 
 - Fix label selector in app values watcher so it supports CAPI clusters.
+- Strip cluster name from App CR name to determine Chart CR name in `chart/current.go` resource to fix WC app updates.
 
 ## [5.7.4] - 2022-03-01